Move the replay device to a global
This way we can create it before the test starts and by the time
we are ready to receive events it will be ready to go. This removes the
need for a lot of the timeouts since we are no longer racing, getting
the Gtk window up and ibus ready takes longer than adding the input
device to the compositor.