Fork off the engine handling into a separate process
With GTK_IM_MODULE=ibus we enter a deadlock where the GTK bits are
dbus-waiting for the engine to reply (our
engine_process_key_event_cb()). But that's in the same process so we
end up with a DBus deadlock.

Fix this by forking immediately on the test and moving the ibus engine
handling to a separate process. This requires gtk_init to be moved to
after the fork, otherwise we hang in ibus_bus_new().

Once our test is complete we send a message to the pipe for the engine
process to shut down.

This is a messy patch because it's partially incomplete
- gtk_init is now missing from test_init, needs to be updated
- the pipefd should ideally be passed around instead of being a global
- the same changes need to be made for the test_keypress_from_recording